completion: avoid panic when completing non-normal paths
Fixes #6861.

`current` might not be a prefix of the (normal) completion candidate
`path` if `current` itself is non-normal. In case `current` is longer
than the candidate `path`, the code previously panicked. The tests have
been extended to trigger this panic.

The panic is avoided by using `strip_prefix()` instead of direct
slicing. However, in case `current` does not prefix `path`, the path
may be truncated too early. Additionally, fish discards any completion
candidates that do not start with `current`, even if the completion
script writes them to stdout.

Thus, this commit is the minimal change to avoid the panic, but it does
not yet handle completion of non-normal paths correctly.