completion: normalize file paths to `RepoPath`
Fixes #7024.

Use `RepoPathBuf` to normalize the current prefix path to forward
slashes. The `path` now also has to be passed as a `&RepoPath`. In some
cases (`log`, `diff`), this can be achieved by omitting `.display()` in
the command's template argument; in the case of `resolve --list`, it is
also normalized via `RepoPathBuf`.

The completion is now obtained from the `RepoPath::as_internal_string`
representation which uses forward slashes on all platforms. That way,
the completion is now also usable in git-bash.exe (Git for Windows with
`COMPLETE=bash`). When using dynamic completion with PowerShell, the
completion results are still valid, as PowerShell can tolerate forward
slashes.