cli: check only present tracking remote refs
Various commands check if the bookmarks they operate on have a tracking
remote ref. If they do, an error or warning is generated. In most of
these cases, it's impossible for the tracked remote ref to be absent.
This includes the checks of:
- `jj git push`
- `jj bookmark set`
- `jj bookmark create`

In the case of `jj abandon`, the generated warning tells users how to
delete the bookmark on the remote. This makes no sense if the remote ref
doesn't even exist. So, for `jj abandon`, it makes sense if the check
only returns `true` if the remote ref is actually present.

The last case where this check happens is `jj bookmark rename`. A
subsequent patch will make rename preserve the tracking state of the old
bookmark for the new one. For that, it is also convenient if the check
ignores absent remote refs. It makes it easier to detect whether the
tracking state of the new bookmark was derived from the old bookmark or
because a tracking bookmark of the same name was previously deleted.